+++ src/blocks/button.js
@@ -17,7 +17,7 @@ const buttonToMjml = (item, blockRenderer, isTopLevelNode) => {
 
     // mj-spacer NOT allowed in mj-body
     return isTopLevelNode
-        ? `<mj-section ><mj-column >${mjButton}</mj-column></mj-section>`
+        ? `<mj-section padding="0px"><mj-column padding="0px">${mjButton}</mj-column></mj-section>`
         : mjButton;
 };